Refactor of querytool.js
