--- src/pages/QueryTool/QueryTool.css
@@ -1,176 +1,0 @@
-.query-tool-container {
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  min-height: calc(100vh - 60px);
-}
-
-/* .title-container{
-    background-color: #302F2F;
-} */
-
-.query-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100%; /* Optional: Ensure it takes full height */
-}
-
-.query-tool {
-  min-width: 80%;
-}
-
-.query-tool .p-tabview-panels .p-tabview-panel {
-  height: 100%; /* Adjusted height to account for header, padding, etc. */
-  /* background-color: green; */
-  flex-grow: 1;
-  overflow: auto;
-
-  /* background-color:#ff0000; */
-}
-
-/* .query-section{
-    background-color: #302F2F;
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-    padding: 15px;
-    margin-bottom: 15px;
-} */
-
-/* .query-input {
-    background-color: #302F2F;
-    color: white;
-    border: 1px solid #302F2F;
-    border-radius: 5px;
-    padding: 10px;
-    width: 100%;
-    height: 100%;
-    display: flex; 
-    align-items: center;
-    gap: 10px; 
-} */
-
-.slider-container {
-  display: flex;
-  align-items: center;
-  justify-content: center; /* This centers the slider */
-  width: 100%;
-  padding: 10px;
-  background-color: #343540;
-  position: relative; /* Allows for absolute positioning of the help icon */
-}
-
-.help-icon {
-  position: absolute;
-  bottom: 10px; /* Adjust as needed */
-  right: 10px; /* Adjust as needed */
-  font-size: 2.5rem;
-  cursor: pointer; /* Change the cursor to indicate it's clickable */
-}
-
-.start-date {
-  width: 100%;
-}
-
-.date {
-  width: 45%;
-  margin-right: 10px;
-  margin-left: 10px;
-}
-
-.spinner-wrapper {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100%; /* Ensure the parent takes up full viewport height */
-  overflow: hidden;
-}
-
-.bottom-right-overlay {
-  position: fixed;
-  bottom: 0;
-  right: 0;
-  margin: 20px;
-  z-index: 1000; /* Ensure it appears above other elements */
-  max-width: 400px; /* Adjust as needed */
-  max-height: 600px;
-}
-
-.query-section {
-  /* background-color: var(--query-color); Use the variable */
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  padding: 15px;
-  margin-bottom: 15px;
-  /* color:black; */
-}
-
-/* .query-input {
-  background-color: #343540;
-  border: #343540;
-} */
-
-.query-input {
-  border-radius: 5px;
-  padding: 10px;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-/* .title-container {
-  background-color: none;
-} */
-
-.title-container h1 {
-  font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-  text-align: center;
-  font-size: 2rem;
-  margin: 0;
-  padding: 10px;
-}
-
-.action-link {
-  text-decoration: none;
-  cursor: pointer;
-}
-
-.action-link:hover {
-  text-decoration: underline;
-}
-
-.show-more {
-  cursor: pointer;
-  margin-left: 5px;
-  text-decoration: none;
-  transition: text-decoration 0.3s, color 0.3s;
-}
-
-.show-more:hover {
-  text-decoration: underline;
-}
-
-.back-button {
-  padding: 11.5px 20px;
-  font-size: 16px;
-  font-style: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-  border-radius: 5px;
-  display: flex;
-  align-items: center;
-  border: none;
-  cursor: pointer;
-}
-
-.left-arrow {
-  margin-right: 10px;
-}
-
-.table-header {
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-}
