+++ src/pages/QueryTool/styles/HelpPanel.css
@@ -1,0 +1,9 @@
+.bottom-right-overlay {
+    position: fixed;
+    bottom: 0;
+    right: 0;
+    margin: 20px;
+    z-index: 1000;
+    max-width: 400px;
+    max-height: 600px;
+  }
