+++ src/pages/QueryTool/styles/QuerySection.css
@@ -1,0 +1,37 @@
+.query-section {
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    padding: 15px;
+    margin-bottom: 15px;
+  }
+  
+  .query-input {
+    border-radius: 5px;
+    padding: 10px;
+    width: 100%;
+    height: 100%;
+    display: flex;
+    align-items: center;
+    gap: 10px;
+  }
+  
+  .slider-container {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    padding: 10px;
+    background-color: #343540;
+    position: relative;
+  }
+  
+  .date {
+    width: 45%;
+    margin-right: 10px;
+    margin-left: 10px;
+  }
+  
+  .start-date {
+    width: 100%;
+  }
