+++ src/pages/QueryTool/styles/QueryTabs.css
@@ -1,0 +1,5 @@
+.query-tool .p-tabview-panels .p-tabview-panel {
+    height: 100%;
+    flex-grow: 1;
+    overflow: auto;
+  }
