+++ src/pages/QueryTool/styles/QueryTool.css
@@ -1,0 +1,33 @@
+.query-tool-container {
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  min-height: calc(100vh - 60px);
+}
+
+.query-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+}
+
+.query-tool {
+  min-width: 80%;
+}
+
+.title-container h1 {
+  font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
+  text-align: center;
+  font-size: 2rem;
+  margin: 0;
+  padding: 10px;
+}
+
+.help-icon {
+  position: absolute;
+  bottom: 10px;
+  right: 10px;
+  font-size: 2.5rem;
+  cursor: pointer;
+}
