+++ src/pages/QueryTool/styles/ResultsTable.css
@@ -1,0 +1,48 @@
+.table-header {
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+  }
+  
+  .spinner-wrapper {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100%;
+    overflow: hidden;
+  }
+  
+  .action-link {
+    text-decoration: none;
+    cursor: pointer;
+  }
+  
+  .action-link:hover {
+    text-decoration: underline;
+  }
+  
+  .show-more {
+    cursor: pointer;
+    margin-left: 5px;
+    text-decoration: none;
+    transition: text-decoration 0.3s, color 0.3s;
+  }
+  
+  .show-more:hover {
+    text-decoration: underline;
+  }
+  
+  .back-button {
+    padding: 11.5px 20px;
+    font-size: 16px;
+    font-style: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
+    border-radius: 5px;
+    display: flex;
+    align-items: center;
+    border: none;
+    cursor: pointer;
+  }
+  
+  .left-arrow {
+    margin-right: 10px;
+  }
